Create display_helper to seperate CLIInterface menu code
