+++ cli/display_helper.rb
@@ -1,0 +1,35 @@
+class DisplayHelper
+  def initialize(api_client)
+    @api_client = api_client
+  end
+
+  def display_team(team)
+    puts "ID: #{team['id']}"
+    puts "Name: #{team['name']}"
+    puts "City: #{team['city']}"
+    puts "Player Count: #{team['players'].count}"
+  end
+
+  def display_player(player)
+    puts "ID: #{player['id']}"
+    puts "Name: #{player['name']}"
+    puts "Number: #{player['number']}"
+    puts "Position: #{player['position']}"
+    puts "Team: #{player['team']['name']}"
+  end
+
+  def display_team_player(player)
+    puts "ID: #{player['id']}  Name: #{player['name']}  Number: #{player['number']}  Position: #{player['position']}"
+  end
+
+  def show_teams_info
+    teams_response = @api_client.show_teams
+    if teams_response.is_a?(Array) && !teams_response.empty?
+      puts "Available Teams:"
+      teams_response.each { |team| puts "#{team['id']}. #{team['name']} || Player Count: #{team['players'].count}" }
+    else
+      puts "No Teams available."
+      nil
+    end
+  end
+end
